requirements.txt
# ---- API & Config ----
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
pydantic>=2.7,<3
python-dotenv>=1.0
pyyaml>=6.0

# ---- Data / Batch ----
pandas>=2.2,<3
pyarrow>=16.0,<18
numpy>=1.26,<3

# ---- Storage / Vector Search ----
pymongo>=4.8,<5
faiss-cpu>=1.8.0   # used by retrieval/vector_store.py

# ---- Scheduling / HTTP utils ----
apscheduler>=3.10,<4
requests>=2.32,<3

# ---- Testing ----
pytest>=8.0,<9

# ---- Embeddings (HF) [optional] ----
# sentence-transformers>=2.7
# torch>=2.2 ; sys_platform != "win32" or platform_machine == "AMD64"
# NOTE: On Windows, install torch using the command from pytorch.org if pip fails.

# ---- Learning-to-Rank (later) ----
# scikit-learn>=1.4
# lightgbm>=4.3     # may require VS Build Tools on Windows

# ---- ONNX serving (for ranker v1) ----
# onnxruntime>=1.18
